refactor: consolidate repository adapters into domain layer
