+++ app/routes/_middleware.ts
@@ -1,7 +1,7 @@
 import { etag } from "hono/etag"
 import { createMiddleware } from "hono/factory"
 import { createRoute } from "honox/factory"
-import { createDbClient } from "../infrastructure/db/client"
+import { createDbClient } from "../libs/db/client"
 
 export default createRoute(
   createMiddleware(async (c, next) => {
